requirements.txt
django
django-parler
djangorestframework
django-parler-rest
django-user-media
Pillow
django-libs
